Move the ASP impl of AppContextFactory
